+++ tests/setup/install_dep.R
@@ -67,12 +67,12 @@ installDep <- function(pkg, devel_mode, immcantation = immcantation_packages,
             # Install from CRAN
             tryCatch(
                 {
-                    devtools::install_version(pkg_name, paste(pkg_logic, pkg_version), repos = "http://lib.stat.cmu.edu/R/CRAN/")
+                    devtools::install_version(pkg_name, paste(pkg_logic, pkg_version), repos = "http://lib.stat.cmu.edu/R/CRAN/", upgrade = "never")
                 },
                 error = function(e) {
                     # This is needed if there is an Immcantation release package that is not
                     # available from CRAN
-                    cat(e, "\n")
+                    cat(as.character(e), "\n")
                     message("Installing ", pkg, " from GitHub...\n ")
                     install_github(paste0("immcantation/", pkg_name, "@", pkg_version), build = TRUE)
                 }
@@ -81,7 +81,7 @@ installDep <- function(pkg, devel_mode, immcantation = immcantation_packages,
     } else {
         if (!document[[pkg_name]]) {
             message(paste0(pkg, ": installing most recent version from GitHub @master."))
-            install_github(paste0("immcantation/", pkg_name, "@master"), upgrade = "never")
+            install_github(paste0("immcantation/", pkg_name, "@master"), upgrade = "never", force = FALSE)
         } else {
             message(paste0(pkg, ": installing and documenting most recent version from GitHub @master."))
             pkg_tmp_dir <- file.path(tempdir(), pkg_name)
@@ -98,10 +98,10 @@ installDep <- function(pkg, devel_mode, immcantation = immcantation_packages,
                     ))
 
                     # Install dependencies, document, build and install
-                    devtools::install_deps(pkg_tmp_dir, dependencies = TRUE, upgrade = "never")
+                    devtools::install_deps(pkg_tmp_dir, dependencies = TRUE, upgrade = "never", force = FALSE)
                     devtools::document(pkg_tmp_dir)
                     devtools::build(pkg_tmp_dir)
-                    devtools::install(pkg_tmp_dir)
+                    devtools::install(pkg_tmp_dir, upgrade = "never", force = FALSE)
                     unlink(pkg_tmp_dir, recursive = TRUE)
                     message(paste0("Successfully installed ", pkg, " from GitHub master"))
                 },
@@ -114,7 +114,7 @@ installDep <- function(pkg, devel_mode, immcantation = immcantation_packages,
 }
 
 # Parse this package version in DESCRIPTION
-this_pkg_version <- read.dcf("DESCRIPTION", fields="Version")
+this_pkg_version <- read.dcf("DESCRIPTION", fields = "Version")
 this_pkg_version <- gsub(".*\\([^0-9.]*(.*)\\)$", "\\1", this_pkg_version)
 
 # If the package is using the devel version number (ends in .999)
@@ -122,7 +122,7 @@ this_pkg_version <- gsub(".*\\([^0-9.]*(.*)\\)$", "\\1", this_pkg_version)
 devel_mode <- grepl("\\.999$", this_pkg_version)
 
 # Parse Imports field in DESCRIPTION
-d <- read.dcf("DESCRIPTION", fields="Imports")
+d <- read.dcf("DESCRIPTION", fields = "Imports")
 d <- sub("^\\n", "", d)
 imports <- strsplit(d, ",\n")[[1]]
 
@@ -132,13 +132,13 @@ imports <- imports[unique(c(immcantation, 1:length(imports)))]
 # Skip bioconductor packages
 imports <- imports[!imports %in% bioconductor_deps]
 
+# Install Bioconductor dependencies first
+if (!is.null(bioconductor_deps)) {
+    BiocManager::install(bioconductor_deps, update=FALSE, ask=FALSE, force=FALSE)
+}
 
 # Install
 for (i in 1:length(imports)) {
     this_import <- imports[i]
     installDep(this_import, devel_mode)
-}
-
-if (!is.null(bioconductor_deps)) {
-    BiocManager::install(bioconductor_deps)
 }